% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFactor-utils.R
\name{combine_lmBF_arrays}
\alias{combine_lmBF_arrays}
\title{A helper function that collapses variables corresponding to arrays in the \code{lmBF} output for neater visuals.}
\usage{
combine_lmBF_arrays(x)
}
\description{
A helper function that collapses variables corresponding to arrays in the \code{lmBF} output for neater visuals.
}
