% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_p.R
\name{my_psinib}
\alias{my_psinib}
\title{Calls sinib::psinib, but handles some extra edge cases}
\usage{
my_psinib(q, size, prob, lower.tail = TRUE, log.p = FALSE)
}
\description{
Calls sinib::psinib, but handles some extra edge cases
}
